slack-rag/slack_bot/requirements.txt
slack_bolt==1.19.0
httpx==0.27.0
python-dotenv==1.0.1
aiohttp==3.8.1